feat: enable strict mode
ref: #17

BREAKING CHANGE: Service hooks could return undefined if consumed in a component that wrapped by the service provider component. This was always the case but not reflected in the types. Now the return type of a hook is the API or undefined. This may cause breaking changes at build time if consuming project has strict typing enabled, and require consumers to check if the result of their hook is defined before using it.